modified:   src/main/java/org/example/psychologicalcounseling/controller/UserInfoController.java 	modified:   src/main/java/org/example/psychologicalcounseling/model/Admin.java 	modified:   src/main/java/org/example/psychologicalcounseling/model/User.java 	new file:   src/main/java/org/example/psychologicalcounseling/module/user/info/EditRequestDto.java 	modified:   src/main/java/org/example/psychologicalcounseling/module/user/info/UserInfoService.java 	deleted:    src/main/java/org/example/psychologicalcounseling/module/user/register/RegisterResponse.java 	deleted:    src/main/java/org/example/psychologicalcounseling/module/user/register/UserRegisterController.java 	deleted:    src/main/java/org/example/psychologicalcounseling/module/user/register/UserRegisterService.java 	modified:   src/main/java/org/example/psychologicalcounseling/repository/AccountRepository.java 	modified:   src/main/java/org/example/psychologicalcounseling/repository/UserRepository.java

+++ src/main/java/org/example/psychologicalcounseling/repository/UserRepository.java
@@ -6,10 +6,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     //表格如同实体类User，uid是主键，主键为Long类型
     //继承了JpaRepository，自带对于Users表格的增删改查方法
 
     User findByEmail(String email);
 
+    //User findById(Long uid);
 }
